Move: CategoryEntity 파일 이동
