+++ user-api/src/main/java/com/biengual/userapi/category/repository/CategoryRepository.java
@@ -3,7 +3,7 @@ package com.biengual.userapi.category.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.biengual.userapi.category.domain.entity.CategoryEntity;
+import com.biengual.userapi.category.domain.CategoryEntity;
 
 import java.util.Optional;
